+++ software/generic/Types.hpp
@@ -16,14 +16,14 @@ uint16_t random(uint16_t max);
 uint16_t random(uint16_t min, uint16_t max);
 
 // Provide the millis function, and allowing setting it for tests.
-uint16_t millis();
-void setMillis(uint16_t time);
-void advanceMillis(uint16_t time);
+uint32_t millis();
+void setMillis(uint32_t time);
+void advanceMillis(uint32_t time);
 
 #else
 // Arduino
 #include "Arduino.h"
-#include "String.h"
+#include "string.h"
 #endif
 
 #endif
